Fix bug causing player's stored Ignores to not load
Skipping the load from database if the player is ignored by someone else
was a logic error. It meant that if the player in question was ignored
by somebody online, then when they loaded we wouldn't load the people
they're ignoring from the database, making it as if they weren't loading
anybody.

The player could then add already-ignored players to end up with
duplicates. We should consider going through the database and removing
duplicates, but for now fix this just by switching to a HashSet instead
of ArrayDeque (which should improve performance anyway.)